pyproject.toml
[project]
name = "gitinspectorgui"
version = "0.4.5"
description = "A CLI and GUI tool for numerical and blame analysis of git repositories."
authors = [{ name = "Bert van Beek", email = "d.a.v.beek@tue.nl" }]
license = { text = "MIT" }
readme = { file = "README.rst", content-type = "text/x-rst" }
keywords = [
    "git",
    "repository",
    "repositories",
    "analyzer",
    "analysis",
    "statistics",
    "blame",
    "cli",
    "gui",
]
classifiers = [
    "Development Status :: 4 - Beta",
    # "Development Status :: 5 - Production/Stable",
    "License :: OSI Approved :: MIT License",
    "Natural Language :: English",
    "Operating System :: MacOS",
    "Operating System :: Microsoft :: Windows :: Windows 10",
    "Operating System :: Microsoft :: Windows :: Windows 11",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3.12",
]
requires-python = ">=3.12,<3.13"
dependencies = [
    "beautifulsoup4 >= 4.12.3",
    "types-beautifulsoup4 >= 4.12",
    "colorlog >= 6.9",
    "gitpython >= 3.1.43",
    "jinja2 >= 3.1.4",
    "jsonschema >= 4.23",
    "types-jsonschema >= 4.23",
    "platformdirs >= 4.3.6",
    "pysimplegui-4-foss == 4.60.4.1",
    "werkzeug >= 3.1.3",
    "xlsxwriter >= 3.2",
    "requests == 2.32.3",                # fixed version for requests and dependencies
    "types-requests == 2.32.0.20241016", #
    "urllib3 == 2.3.0",                  # required by requests
    "charset-normalizer == 3.4.1",       # required by requests
    "ruff>=0.9.7",
    "dmgbuild>=1.6.4",
]

[project.urls]
homepage = "https://gitinspectorgui.readthedocs.io/en/latest/index.html"
repository = "https://github.com/davbeek/gitinspectorgui"

[project.scripts]
gigui = "gigui.cli:main"
gitinspectorgui = "gigui.cli:main"

[dependency-groups]
dev = [
    "pre-commit >= 4.0.1",
    "pyinstaller >= 6.11.1",
    "ruff >= 0.9",
    "sphinx >=7,<9",
]

[tool.hatch.build.targets.wheel]
packages = ["src/gigui"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
index-strategy = "unsafe-best-match"

[[tool.uv.index]]
url = "https://pypi.org/simple"
default = true

[[tool.uv.index]]
name = "testpypi"
url = "https://test.pypi.org/simple/"
publish-url = "https://test.pypi.org/legacy/"
